requirements.txt
imageio==2.9.0
ipython==7.18.1
Keras==2.4.3
matplotlib==3.2.2
numpy==1.18.5
Pillow==7.2.0
tensorflow==2.3.0
